+++ src/main/java/jp/ac/titech/c/se/stein/core/ConcurrentRepositoryRewriter.java
@@ -1,4 +1,4 @@
-package finergit.rewrite;
+package jp.ac.titech.c.se.stein.core;
 
 import java.util.HashMap;
 import java.util.Spliterator;
@@ -14,7 +14,7 @@ import org.eclipse.jgit.revwalk.RevWalk;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import finergit.rewrite.Try.ThrowableFunction;
+import jp.ac.titech.c.se.stein.core.Try.ThrowableFunction;
 
 public class ConcurrentRepositoryRewriter extends RepositoryRewriter {
     private static final Logger log = LoggerFactory.getLogger(ConcurrentRepositoryRewriter.class);
