+++ src/main/java/jp/ac/titech/c/se/stein/core/EntrySet.java
@@ -1,4 +1,4 @@
-package finergit.rewrite;
+package jp.ac.titech.c.se.stein.core;
 
 import java.util.ArrayList;
 import java.util.Collections;
