+++ src/main/java/jp/ac/titech/c/se/stein/core/Try.java
@@ -1,4 +1,4 @@
-package finergit.rewrite;
+package jp.ac.titech.c.se.stein.core;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
